+++ include/fiction/physical_design/surface_analysis.hpp
@@ -22,7 +22,6 @@
 #include "fiction/technology/fcn/cell_ports.hpp"
 #include "fiction/technology/sidb/lattice.hpp"
 #include "fiction/technology/sidb/layout.hpp"
-#include "fiction/technology/sidb/technology.hpp"
 #include "fiction/traits.hpp"
 
 #include <kitty/dynamic_truth_table.hpp>