+++ include/fiction/technology/sidb/io/read_sqd_layout.hpp
@@ -21,7 +21,6 @@
 #include "fiction/technology/sidb/lattice.hpp"
 #include "fiction/technology/sidb/layout.hpp"
 #include "fiction/technology/sidb/model/defect.hpp"
-#include "fiction/technology/sidb/technology.hpp"
 
 #include <tinyxml2.h>