+++ include/fiction/technology/sidb/simulation/engines/quickexact.hpp
@@ -28,7 +28,6 @@
 #include "fiction/technology/sidb/simulation/detail/simulation_state.hpp"
 #include "fiction/technology/sidb/simulation/potential_landscape.hpp"
 #include "fiction/technology/sidb/simulation/result.hpp"
-#include "fiction/technology/sidb/technology.hpp"
 #include "fiction/utils/execution_timeout.hpp"
 #include "fiction/utils/math/gray_code_iterator.hpp"
 #include "fiction/utils/progress.hpp"