+++ test/technology/sidb/simulation/analysis/test_can_positive_charges_occur.cpp
@@ -21,7 +21,6 @@
 #include <fiction/technology/sidb/layout.hpp>
 #include <fiction/technology/sidb/model/simulation_parameters.hpp>
 #include <fiction/technology/sidb/simulation/analysis/can_positive_charges_occur.hpp>
-#include <fiction/technology/sidb/technology.hpp>
 
 #include <stdexcept>